auth/tests/functional/requirements.txt
pytest==6.1.2
pytest-asyncio==0.12.0
aiohttp==3.7.2
asyncpg==0.27.0
psycopg2-binary==2.9.3
redis==3.5.3